# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/torrent/add": {
      "post": {
        "tags": [
          "Downloads"
        ],
        "summary": "Add a torrent to the list",
        "description": "",
        "parameters": [
          {
            "name": "torrent",
            "in": "body",
            "description": "Add a torrent",
            "schema": {
              "type": "object",
              "properties": {
                "magnet": {
                  "type": "string",
                  "example": "magnet:?xt=urn:btih:08ada5a7a6183aae1e09d831df6748d566095a10&dn=Sintel&tr=udp%3A%2F%2Fexplodie.org%3A6969&tr=udp%3A%2F%2Ftracker.coppersurfer.tk%3A6969&tr=udp%3A%2F%2Ftracker.empire-js.us%3A1337&tr=udp%3A%2F%2Ftracker.leechers-paradise.org%3A6969&tr=udp%3A%2F%2Ftracker.opentrackr.org%3A1337&tr=wss%3A%2F%2Ftracker.btorrent.xyz&tr=wss%3A%2F%2Ftracker.fastcast.nz&tr=wss%3A%2F%2Ftracker.openwebtorrent.com&ws=https%3A%2F%2Fwebtorrent.io%2Ftorrents%2F&xs=https%3A%2F%2Fwebtorrent.io%2Ftorrents%2Fsintel.torrent"
                }
              },
              "required": [
                "magnet"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "418": {
            "description": "I'm a teapot"
          }
        }
      }
    },
    "/torrent/pause": {
      "post": {
        "tags": [
          "Downloads"
        ],
        "summary": "Pause a torrent in the list",
        "description": "",
        "parameters": [
          {
            "name": "torrent",
            "in": "body",
            "description": "Pause a torrent",
            "schema": {
              "type": "object",
              "properties": {
                "magnet": {
                  "type": "string",
                  "example": "magnet:?xt=urn:btih:08ada5a7a6183aae1e09d831df6748d566095a10&dn=Sintelbuild&tr=udp%3A%2F%2Fexplodie.org%3A6969&tr=udp%3A%2F%2Ftracker.coppersurfer.tk%3A6969&tr=udp%3A%2F%2Ftracker.empire-js.us%3A1337&tr=udp%3A%2F%2Ftracker.leechers-paradise.org%3A6969&tr=udp%3A%2F%2Ftracker.opentrackr.org%3A1337&tr=wss%3A%2F%2Ftracker.btorrent.xyz&tr=wss%3A%2F%2Ftracker.fastcast.nz&tr=wss%3A%2F%2Ftracker.openwebtorrent.com&ws=https%3A%2F%2Fwebtorrent.io%2Ftorrents%2F&xs=https%3A%2F%2Fwebtorrent.io%2Ftorrents%2Fsintel.torrent"
                }
              },
              "required": [
                "magnet"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A single torrent information",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "magnet": {
                    "type": "string",
                    "example": "magnet:?xt=urn:btih:08ada5a7a6183aae1e09d831df6748d566095a10&dn=Sintel&tr=udp%3A%2F%2Fexplodie.org%3A6969&tr=udp%3A%2F%2Ftracker.coppersurfer.tk%3A6969&tr=udp%3A%2F%2Ftracker.empire-js.us%3A1337&tr=udp%3A%2F%2Ftracker.leechers-paradise.org%3A6969&tr=udp%3A%2F%2Ftracker.opentrackr.org%3A1337&tr=wss%3A%2F%2Ftracker.btorrent.xyz&tr=wss%3A%2F%2Ftracker.fastcast.nz&tr=wss%3A%2F%2Ftracker.openwebtorrent.com&ws=https%3A%2F%2Fwebtorrent.io%2Ftorrents%2F&xs=https%3A%2F%2Fwebtorrent.io%2Ftorrents%2Fsintel.torrent"
                  }
                },
                "required": [
                  "magnet"
                ]
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/torrent/check-status": {
      "get": {
        "tags": [
          "Downloads"
        ],
        "summary": "Check the status of the torrents",
        "description": "It returns all the detail about the torrent",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Downloads"
        ],
        "summary": "Check the status of a single torrent",
        "description": "It returns all the detail about the torrent",
        "parameters": [
          {
            "name": "torrent",
            "in": "body",
            "description": "Id of the torrent",
            "schema": {
              "type": "object",
              "properties": {
                "magnet": {
                  "type": "string",
                  "example": "magnet:?xt=urn:btih:08ada5a7a6183aae1e09d831df6748d566095a10&dn=Sintel&tr=udp%3A%2F%2Fexplodie.org%3A6969&tr=udp%3A%2F%2Ftracker.coppersurfer.tk%3A6969&tr=udp%3A%2F%2Ftracker.empire-js.us%3A1337&tr=udp%3A%2F%2Ftracker.leechers-paradise.org%3A6969&tr=udp%3A%2F%2Ftracker.opentrackr.org%3A1337&tr=wss%3A%2F%2Ftracker.btorrent.xyz&tr=wss%3A%2F%2Ftracker.fastcast.nz&tr=wss%3A%2F%2Ftracker.openwebtorrent.com&ws=https%3A%2F%2Fwebtorrent.io%2Ftorrents%2F&xs=https%3A%2F%2Fwebtorrent.io%2Ftorrents%2Fsintel.torrent"
                }
              },
              "required": [
                "magnet"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/torrent/get-file/{filename}": {
      "get": {
        "tags": [
          "Downloads"
        ],
        "summary": "Return the torrent file of selected value",
        "description": "From the infoHash return the file to download",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "torrentId",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "The torrent file"
          }
        }
      }
    },
    "/torrent/destroy": {
      "post": {
        "tags": [
          "Downloads"
        ],
        "summary": "Remove a torrent in the list",
        "description": "",
        "parameters": [
          {
            "name": "torrent",
            "in": "body",
            "description": "Remove a torrent and destroy data",
            "schema": {
              "type": "object",
              "properties": {
                "magnet": {
                  "type": "string",
                  "example": "magnet:?xt=urn:btih:08ada5a7a6183aae1e09d831df6748d566095a10&dn=Sintel&tr=udp%3A%2F%2Fexplodie.org%3A6969&tr=udp%3A%2F%2Ftracker.coppersurfer.tk%3A6969&tr=udp%3A%2F%2Ftracker.empire-js.us%3A1337&tr=udp%3A%2F%2Ftracker.leechers-paradise.org%3A6969&tr=udp%3A%2F%2Ftracker.opentrackr.org%3A1337&tr=wss%3A%2F%2Ftracker.btorrent.xyz&tr=wss%3A%2F%2Ftracker.fastcast.nz&tr=wss%3A%2F%2Ftracker.openwebtorrent.com&ws=https%3A%2F%2Fwebtorrent.io%2Ftorrents%2F&xs=https%3A%2F%2Fwebtorrent.io%2Ftorrents%2Fsintel.torrent"
                }
              },
              "required": [
                "magnet"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A single torrent information",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "magnet": {
                    "type": "string",
                    "example": "magnet:?xt=urn:btih:08ada5a7a6183aae1e09d831df6748d566095a10&dn=Sintel&tr=udp%3A%2F%2Fexplodie.org%3A6969&tr=udp%3A%2F%2Ftracker.coppersurfer.tk%3A6969&tr=udp%3A%2F%2Ftracker.empire-js.us%3A1337&tr=udp%3A%2F%2Ftracker.leechers-paradise.org%3A6969&tr=udp%3A%2F%2Ftracker.opentrackr.org%3A1337&tr=wss%3A%2F%2Ftracker.btorrent.xyz&tr=wss%3A%2F%2Ftracker.fastcast.nz&tr=wss%3A%2F%2Ftracker.openwebtorrent.com&ws=https%3A%2F%2Fwebtorrent.io%2Ftorrents%2F&xs=https%3A%2F%2Fwebtorrent.io%2Ftorrents%2Fsintel.torrent"
                  }
                },
                "required": [
                  "magnet"
                ]
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/torrent/remove": {
      "post": {
        "tags": [
          "Downloads"
        ],
        "summary": "Remove a torrent in the list, keeping the data",
        "description": "",
        "parameters": [
          {
            "name": "torrent",
            "in": "body",
            "description": "Remove a torrent and keep data",
            "schema": {
              "type": "object",
              "properties": {
                "magnet": {
                  "type": "string",
                  "example": "magnet:?xt=urn:btih:08ada5a7a6183aae1e09d831df6748d566095a10&dn=Sintel&tr=udp%3A%2F%2Fexplodie.org%3A6969&tr=udp%3A%2F%2Ftracker.coppersurfer.tk%3A6969&tr=udp%3A%2F%2Ftracker.empire-js.us%3A1337&tr=udp%3A%2F%2Ftracker.leechers-paradise.org%3A6969&tr=udp%3A%2F%2Ftracker.opentrackr.org%3A1337&tr=wss%3A%2F%2Ftracker.btorrent.xyz&tr=wss%3A%2F%2Ftracker.fastcast.nz&tr=wss%3A%2F%2Ftracker.openwebtorrent.com&ws=https%3A%2F%2Fwebtorrent.io%2Ftorrents%2F&xs=https%3A%2F%2Fwebtorrent.io%2Ftorrents%2Fsintel.torrent"
                }
              },
              "required": [
                "magnet"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A single torrent information",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "magnet": {
                    "type": "string",
                    "example": "magnet:?xt=urn:btih:08ada5a7a6183aae1e09d831df6748d566095a10&dn=Sintel&tr=udp%3A%2F%2Fexplodie.org%3A6969&tr=udp%3A%2F%2Ftracker.coppersurfer.tk%3A6969&tr=udp%3A%2F%2Ftracker.empire-js.us%3A1337&tr=udp%3A%2F%2Ftracker.leechers-paradise.org%3A6969&tr=udp%3A%2F%2Ftracker.opentrackr.org%3A1337&tr=wss%3A%2F%2Ftracker.btorrent.xyz&tr=wss%3A%2F%2Ftracker.fastcast.nz&tr=wss%3A%2F%2Ftracker.openwebtorrent.com&ws=https%3A%2F%2Fwebtorrent.io%2Ftorrents%2F&xs=https%3A%2F%2Fwebtorrent.io%2Ftorrents%2Fsintel.torrent"
                  }
                },
                "required": [
                  "magnet"
                ]
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  }
}